Se ajusta en configuraciones, la dinamica, el radio de busqueda, se añande la fecha deregistro a la pagina de usuarios

+++ lib/models/prices_model.dart
@@ -17,7 +17,7 @@ class Price {
   int theDistanciaTarifaMinima;
   int theNumeroCancelacionesConductor;
   int theNumeroCancelacionesUsuario;
-  int theRadioDeBusqueda;
+  double theRadioDeBusqueda;
   int theRecargaInicial;
   int theTarifaAeropuerto;
   int theTarifaMinimaRegular;
